Maturity Matrix
Markdown conversion of PDF file of Organisation Level Maturity Matrix.

I would appreciate somebody checking the last part under: "Other categories/ criteria to be included" as the lines between cells are not clear and I am not sure if I included the correct items under each of the headings for each Level.

I also didn't include the text under the table as I was not sure how to include it.